[Naming] Rename the rule to make it correspond with what rule does
