+++ config/set/naming.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Rector\Naming\Rector\Assign\RenameVariableToMatchGetMethodNameRector;
+use Rector\Naming\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector;
 use Rector\Naming\Rector\Class_\RenamePropertyToMatchTypeRector;
 use Rector\Naming\Rector\ClassMethod\MakeGetterClassMethodNameStartWithGetRector;
 use Rector\Naming\Rector\ClassMethod\MakeIsserClassMethodNameStartWithIsRector;
@@ -15,7 +15,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     $services->set(RenameParamToMatchTypeRector::class);
     $services->set(RenamePropertyToMatchTypeRector::class);
     $services->set(RenameVariableToMatchNewTypeRector::class);
-    $services->set(RenameVariableToMatchGetMethodNameRector::class);
+    $services->set(RenameVariableToMatchMethodCallReturnTypeRector::class);
     $services->set(MakeGetterClassMethodNameStartWithGetRector::class);
     $services->set(MakeIsserClassMethodNameStartWithIsRector::class);
 };
