--- rules/naming/tests/Rector/Assign/RenameVariableToMatchGetMethodNameRector/Fixture/function.php.inc
@@ -1,43 +1,0 @@
-<?php
-
-namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Fixture;
-
-use Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Source\FastRunner;
-
-function run()
-{
-    $a = getFastRunner();
-    $a->exit();
-}
-
-/**
- * @return FastRunner
- */
-function getFastRunner()
-{
-    return new FastRunner();
-}
-
-?>
------
-<?php
-
-namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Fixture;
-
-use Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Source\FastRunner;
-
-function run()
-{
-    $fastRunner = getFastRunner();
-    $fastRunner->exit();
-}
-
-/**
- * @return FastRunner
- */
-function getFastRunner()
-{
-    return new FastRunner();
-}
-
-?>
