--- rules/naming/tests/Rector/Assign/RenameVariableToMatchGetMethodNameRector/Fixture/skip_dynamic_name.php.inc
@@ -1,14 +1,0 @@
-<?php
-
-namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Fixture;
-
-use Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Source\FastRunner;
-
-class SkipDynamicName
-{
-    public function run($name)
-    {
-        /** @var FastRunner $a */
-        $a = $this->{$name->callThat}()();
-    }
-}
