--- rules/naming/tests/Rector/Assign/RenameVariableToMatchGetMethodNameRector/Source/FastRunner.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchGetMethodNameRector\Source;
-
-final class FastRunner
-{
-
-}
