+++ rules/naming/tests/Rector/Assign/RenameVariableToMatchMethodCallReturnTypeRector/Fixture/function.php.inc
@@ -1,0 +1,43 @@
+<?php
+
+namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector\Fixture;
+
+use Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector\Source\FastRunner;
+
+function run()
+{
+    $a = getFastRunner();
+    $a->exit();
+}
+
+/**
+ * @return FastRunner
+ */
+function getFastRunner()
+{
+    return new FastRunner();
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector\Fixture;
+
+use Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector\Source\FastRunner;
+
+function run()
+{
+    $fastRunner = getFastRunner();
+    $fastRunner->exit();
+}
+
+/**
+ * @return FastRunner
+ */
+function getFastRunner()
+{
+    return new FastRunner();
+}
+
+?>
