+++ rules/naming/tests/Rector/Assign/RenameVariableToMatchMethodCallReturnTypeRector/Fixture/skip_dynamic_name.php.inc
@@ -1,0 +1,14 @@
+<?php
+
+namespace Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector\Fixture;
+
+use Rector\Naming\Tests\Rector\Assign\RenameVariableToMatchMethodCallReturnTypeRector\Source\FastRunner;
+
+class SkipDynamicName
+{
+    public function run($name)
+    {
+        /** @var FastRunner $a */
+        $a = $this->{$name->callThat}()();
+    }
+}
